chore: restructure src/views/ into domain-feature subfolders
Group each domain's provider and tree items together:
- views/integrations/ — IntegrationsProvider + Integration, File, Folder, Route
- views/deployments/  — DeploymentsProvider + ChildItem, ParentItem, RootItem, Route
- views/tests/        — TestsProvider + Test, TestFolder
- views/openapi/      — OpenApiProvider + OpenApiFile, OpenApiFolder
- views/infrastructure/ — InfrastructureProvider, RefreshManager, ServiceManager, InfrastructureItem
- views/help/         — HelpFeedbackProvider
- views/shared/       — AbstractFolderTreeProvider, AbstractFolder

Remove emptied folders: providers/, *TreeItems/, treeItems/
Update all import paths in extension, commands, executors, and test files.
